Changement calculerScore
* création constructeur et getters

* ajout 2 ajouter + création calculerSport

* ajout paramètre dans Sport calculerScore

* implémentation début Athletisme

* Implémentation Handball

* Implémentation Natation

* Implémentation Escrime

* Implémentation Volley-ball

* modif calculerScore retour = double

---------

+++ Athletisme.java
@@ -5,7 +5,7 @@ public class Athletisme extends Sport{
     }
     
     @Override
-    public int calculerScore(Athlete athlete) {
+    public double calculerScore(Athlete athlete) {
         return 0.5*athlete.getForce() + athlete.getAgilite() + 2*athlete.getEndurance();
     }
 }
